# docs/swagger.yaml
basePath: /
definitions:
...
  ds.ErrorResponse:
    properties:
      error:
        example: Некорректный токен
        type: string
    type: object
  ds.LoginReq:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  ds.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        type: string
      expires_in:
        description: время жизни токена (в секундах)
        example: 3600
        type: integer
      token_type:
        example: Bearer
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API сервиса аутентификации с Redis и JWT. Поддерживает роли пользователей
    и администраторов.
  title: Device emission Service API
  version: "1.0"
paths:
...
  /api/emissions_calculation/:order_id/device/:material_id:
    delete:
      description: Удаляет устройство из расчета выбросов по ID заказа и ID устройства
      parameters:
      - description: ID заказа
        in: path
        name: order_id
        required: true
        type: integer
      - description: ID устройства (material_id)
        in: path
        name: material_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Устройство успешно удалено из расчета
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID заказа или устройства
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при удалении устройства
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить устройство из заказа
      tags:
      - Заявки с устройствами
...
  /api/emissions_calculation/devices/:order_id/:material_id/custom_power:
    put:
      consumes:
      - application/json
      description: Обновляет значение пользовательской мощности (custom_power) для
        устройства в заказе
      parameters:
      - description: ID заказа
        in: path
        name: order_id
        required: true
        type: integer
      - description: ID устройства (material_id)
        in: path
        name: material_id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Мощность успешно обновлена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID или тело запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при обновлении мощности
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить пользовательскую мощность устройства
      tags:
      - Заявки с устройствами
...
  /api/users/login:
    post:
      consumes:
      - application/json
      description: Авторизует пользователя по логину и паролю, возвращает JWT-токен
        и время жизни
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.LoginResponse'
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Пользователь не найден или неверный пароль
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Аутентификация пользователя
      tags:
      - auth
  /api/users/logout:
    post:
      description: Добавляет JWT-токен в черный список (блеклист) Redis, чтобы он
        стал недействительным
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
        "400":
          description: Некорректный токен
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
        "500":
          description: Ошибка при записи токена в блеклист
          schema:
            $ref: '#/definitions/ds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Деаутентификация пользователя
      tags:
      - auth
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"